+++ vite-env.d.ts
@@ -7,3 +7,8 @@ declare module "*.svg" {
   const src: string;
   export default src;
 }
+
+declare module "*.md?raw" {
+  const content: string;
+  export default content;
+}